Move FLootLockerResponse to it's own file instead of config.h
