+++ LootLockerSDK/Source/LootLockerSDK/Public/GameAPI/LootLockerMessagesRequestHandler.h
@@ -4,7 +4,7 @@
 
 #include "CoreMinimal.h"
 #include "UObject/NoExportTypes.h"
-#include "LootLockerConfig.h"
+#include "LootLockerResponse.h"
 #include "JsonObjectConverter.h"
 #include "LootLockerHttpClient.h"
 #include "LootLockerMessagesRequestHandler.generated.h"
