+++ LootLockerSDK/Source/LootLockerSDK/Public/GameAPI/LootLockerPersistentStorageRequestHandler.h
@@ -3,8 +3,7 @@
 #pragma once
 
 #include "CoreMinimal.h"
-#include "UObject/NoExportTypes.h"
-#include "LootLockerConfig.h"
+#include "LootLockerResponse.h"
 #include "LootLockerHttpClient.h"
 #include "JsonObjectConverter.h"
 #include "LootLockerPersistentStorageRequestHandler.generated.h"
