+++ LootLockerSDK/Source/LootLockerSDK/Public/GameAPI/LootLockerPlayerFilesRequestHandler.h
@@ -3,7 +3,7 @@
 #pragma once
 #include "CoreMinimal.h"
 #include "UObject/NoExportTypes.h"
-#include "LootLockerConfig.h"
+#include "LootLockerResponse.h"
 #include "LootLockerHttpClient.h"
 #include "LootLockerPlayerFilesRequestHandler.generated.h"
