+++ LootLockerSDK/Source/LootLockerSDK/Public/GameAPI/LootLockerPlayerRequestHandler.h
@@ -4,6 +4,7 @@
 
 #include "CoreMinimal.h"
 #include "UObject/NoExportTypes.h"
+#include "LootLockerResponse.h"
 #include "LootLockerConfig.h"
 #include "JsonObjectConverter.h"
 #include "GameAPI/LootLockerAuthenticationRequestHandler.h"
