+++ LootLockerSDK/Source/LootLockerSDK/Public/GameAPI/LootLockerTriggerEventsRequestHandler.h
@@ -4,7 +4,7 @@
 
 #include "CoreMinimal.h"
 #include "UObject/NoExportTypes.h"
-#include "LootLockerConfig.h"
+#include "LootLockerResponse.h"
 #include "JsonObjectConverter.h"
 #include "LootLockerHttpClient.h"
 #include "GameAPI/LootLockerPlayerRequestHandler.h"
