+++ LootLockerSDK/Source/LootLockerSDK/Public/GameAPI/LootLockerUserGeneratedContentRequestHandler.h
@@ -5,6 +5,7 @@
 #include "CoreMinimal.h"
 #include "UObject/NoExportTypes.h"
 #include "LootLockerConfig.h"
+#include "LootLockerResponse.h"
 #include "JsonObjectConverter.h"
 #include "LootLockerHttpClient.h"
 #include "LootLockerUserGeneratedContentRequestHandler.generated.h"
