+++ LootLockerSDK/Source/LootLockerSDK/Public/LootLockerConfig.h
@@ -9,29 +9,6 @@
 
 DECLARE_LOG_CATEGORY_EXTERN(LogLootLockerGameSDK, Log, All);
 
-USTRUCT(BlueprintType)
-struct FLootLockerResponse
-{
-	GENERATED_BODY()
-public:
-	UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
-	bool success = false;
-	UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
-	bool ServerCallHasError = false;
-	UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
-	int ServerCallStatusCode = 0;
-	UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
-	FString FullTextFromServer;
-	UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
-	FString ServerError;
-	UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
-	FString message;
-	UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
-	FString error;
-	UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
-	FString session_token;
-};
-
 UENUM(BlueprintType)
 enum class ELootLockerHTTPMethod : uint8
 {
