+++ LootLockerSDK/Source/LootLockerSDK/Public/LootLockerResponse.h
@@ -1,0 +1,28 @@
+// Copyright (c) 2021 LootLocker
+
+#pragma once
+
+#include "LootLockerResponse.generated.h"
+
+USTRUCT(BlueprintType)
+struct FLootLockerResponse
+{
+	GENERATED_BODY()
+public:
+    UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
+    bool success = false;
+    UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
+    bool ServerCallHasError = false;
+    UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
+    int ServerCallStatusCode = 0;
+    UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
+    FString FullTextFromServer;
+    UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
+    FString ServerError;
+    UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
+    FString message;
+    UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
+    FString error;
+    UPROPERTY(BlueprintReadWrite, EditAnywhere, BlueprintReadWrite, Category = "LootLocker")
+    FString session_token;
+};
